refactor: relocate ContentBlockHelper from Infrastructure/Dto to Domain/Utils
ContentBlockHelper is a utility class with static factory methods for creating
MCP content block DTOs - not a DTO itself, and not infrastructure code. Moving
it to Domain/Utils places it alongside related MCP utilities (McpValidator,
McpAnnotationMapper, SchemaTransformer) where it semantically belongs.

This change also eliminates the single-file folder smell: Infrastructure/Dto
contained only this one class, suggesting a misplaced abstraction.

Changes:
- Move includes/Infrastructure/Dto/ContentBlockHelper.php → Domain/Utils/
- Update namespace from WP\MCP\Infrastructure\Dto to WP\MCP\Domain\Utils
- Update imports in ToolsHandler.php and DtoSerializationRegressionTest.php
- Move test file to tests/Unit/Domain/Utils/ContentBlockHelperTest.php
- Remove empty Infrastructure/Dto directories
