+++ includes/Handlers/Tools/ToolsHandler.php
@@ -11,7 +11,7 @@ namespace WP\MCP\Handlers\Tools;
 
 use WP\MCP\Core\McpServer;
 use WP\MCP\Handlers\HandlerHelperTrait;
-use WP\MCP\Infrastructure\Dto\ContentBlockHelper;
+use WP\MCP\Domain\Utils\ContentBlockHelper;
 use WP\MCP\Infrastructure\ErrorHandling\McpErrorFactory;
 use WP\McpSchema\Server\Tools\CallToolResult;
 use WP\McpSchema\Server\Tools\ListToolsResult;
